Build RPATH wrapper for executables
This puts a wrapper script at $BUILD/swanson, which sets LD_LIBRARY_PATH
appropriately to be able to load in libswanson.so while running swanson.
(CMake solves this by added an actual -rpath to the executable, and then
stripping that rpath when installing.  We can look at doing something
similar, but it will require having chrpath [1] or patchelf [2]
available at install time.)

[1] http://directory.fsf.org/project/chrpath/
[2] http://nixos.org/patchelf.html